+++ CHANGELOG.md
@@ -5,6 +5,13 @@ file is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/), and
 this project adheres to [Semantic
 Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [Unreleased]
+
+### Added
+
+* `RequestCode::AllowList` and `RequestCode::BlockList` for updating
+   allow/block list.
+
 ## [0.7.0] - 2023-05-11
 
 ### Changed